requirements.txt
Django==3.0.1
psycopg2==2.8.4
python-decouple==3.3
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
django-cors-headers==3.2.0
djangorestframework_simplejwt==4.3.0
